requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
pymysql==1.1.0
aiomysql==0.2.0
sqlalchemy==2.0.23
pydantic==2.7.2
pydantic-settings==2.5.2
python-dotenv==1.1.0
loguru==0.7.2
jinja2==3.1.2
python-jose==3.3.0
cryptography==41.0.5
passlib==1.7.4
bcrypt==4.0.1
fastmcp==2.2.5
python-multipart==0.0.6
psutil==5.9.6
requests==2.31.0
uv>=0.7.1
email-validator>=2.2.0
dnspython>=2.7.0

# mysql查询
mcp>=0.1.0
mysql-connector-python==8.0.33
python-dotenv>=0.19.0
# protobuf需要固定版本，解决依赖冲突
protobuf<=3.20.3

# 百度AI搜索
sseclient-py>=1.7.2

# 思维导图
mindmap-mcp-server>=0.1.1

# SQL Server依赖
pyodbc>=4.0.30

